fix(metrics): make the aggregates match the real databases
The end-to-end verification against the running services found three things the
unit tests could not, because they only appear against a real database and a
real gateway.

1. PostgreSQL returns `bigint` from `count(*)`, and Dapper requires a record's
   constructor to match the column types exactly. With `int` in the DTO and
   `bigint` in the query, materialization throws and the request ends in 500.
   That is why /panel and /zonas failed while /rendimiento and /actividad
   worked: their row type is a class with settable properties, which Dapper
   fills more leniently. The counting queries now cast with `::int`, and the
   DTOs record why the cast stays — removing it produces a 500, not a
   compile error.

2. Identity signs with its own issuer, audience and secret
   (`Deuna.Identity` / `Deuna.Api`), not with the shared default. The service
   was validating against the default, so a valid ADMIN token came back as 401
   while a hand-made token signed with the default was accepted. The values now
   come from configuration, like in the rest of the services.

3. The gateway's real destinations are set through environment variables in
   compose (container names, `http://orders:8080`), not through the addresses
   in `appsettings.json`, which are the standalone-run fallbacks. A cluster
   with no variable falls back to `http://localhost:5006`, and inside the
   gateway container `localhost` is the gateway itself, so the route answered
   502. The metrics cluster now has its variable.

Verified: 29/29 checks against the running system. Every number the API returns
was compared against a direct SQL query on the database it comes from — orders,
riders, restaurants and the cross-domain count of restaurants with orders
today. The zone normalization was proven with real data: two orders created
with "Bogotá" and "Bogota" come back as a single zone holding both.

+++ Deuna.Metrics.Service/Services/MetricasDomiciliarios.cs
@@ -45,10 +45,10 @@ public sealed class MetricasDomiciliarios(
                   and "Estado" not in ('Completado', 'Rechazado')
             )
             select
-                count(*) filter (where r."Activo")                                   as Activos,
-                count(*) filter (where r."Activo" and a."RepartidorId" is null)      as Disponibles,
-                count(*) filter (where r."Activo" and a."RepartidorId" is not null)  as EnServicio,
-                count(*) filter (where not r."Activo")                               as EnDescanso
+                count(*) filter (where r."Activo")::int                                   as Activos,
+                count(*) filter (where r."Activo" and a."RepartidorId" is null)::int      as Disponibles,
+                count(*) filter (where r."Activo" and a."RepartidorId" is not null)::int  as EnServicio,
+                count(*) filter (where not r."Activo")::int                               as EnDescanso
             from repartidores_replicados r
             left join abiertas a on a."RepartidorId" = r."Id"
             """;
